Revert "shrink ch^2 diff with direct calc to 4426 - 4426 == -0.000145"
This reverts commit c6166cd678669d84f057529cc4df4467969c141f.

+++ src/GapsParameters.h
@@ -45,10 +45,10 @@ public:
     unsigned outputFrequency;
     unsigned checkpointInterval;
     unsigned snapshotFrequency;
-    double alphaA;
-    double alphaP;
-    double maxGibbsMassA;
-    double maxGibbsMassP;
+    float alphaA;
+    float alphaP;
+    float maxGibbsMassA;
+    float maxGibbsMassP;
     PumpThreshold pumpThreshold;
     GapsAlgorithmPhase snapshotPhase;
     bool useFixedPatterns;
